+++ Parser.kt
@@ -3,7 +3,7 @@ package converter
 class Parser {
     fun parse(input: String): Measure {
         val (number, unit) = input.split(Regex(" "))
-        return try { Measure(number.toInt(), Units.getUnit(unit)) } catch (e: RuntimeException) {
+        return try { Measure(number.toDouble(), Units.getUnit(unit)) } catch (e: RuntimeException) {
             throw RuntimeException(e.message)
         }
     }
